Move enum class FrameStage to FrameStage.h
