+++ Osiris/Hacks/Visuals.h
@@ -3,17 +3,7 @@
 #include <string_view>
 
 #include "../Memory.h"
-
-enum class FrameStage {
-    UNDEFINED = -1,
-    START,
-    NET_UPDATE_START,
-    NET_UPDATE_POSTDATAUPDATE_START,
-    NET_UPDATE_POSTDATAUPDATE_END,
-    NET_UPDATE_END,
-    RENDER_START,
-    RENDER_END
-};
+#include "../SDK/FrameStage.h"
 
 namespace Visuals {
     void colorWorld() noexcept;
